Added a check to see if the config-version property is equal to 1.0.
This defaults to 0.9, that way if you don't have the property in the config then it will be added and the next check will be equal to 1.0. The purpose of this is stop adding the defaults, granted this could probably be done a better way so I am considering this a hotfix for issue #28. No promises this will work, but hey it is a start