Closes #50.
New syntax is similar to Essentials' give command:
    give <item id|item id:damage value> <amount> [item meta...]
where item meta can be one or more of the following:
    <enchantment id>:<enchantment level>
    name:<item name>
    lore:<Item description>

For name and description, use an underscore '_' to create a space. For description only, use a pipe '|' to create a new line. The '&<color id>' syntax can be used to color a meta information.