Moving to 3.0.0, as I want to do some restructuring
Right now the types are stored in the config.yml which is not ideal
because if we want to add new configuration options we have to manually
insert them instead of letting Bukkit automatically add the missing
nodes. I will be moving the types over to types.yml or maybe even store
them all in a types folder which will allow easy sharing of types. As
this is a massive change, I will be taking this opportunity to rework
the parsing of the types and get away from numbers as ids and use the
material names. The internal structure is also changing, moving to a
more class based setup instead of parsing strings all of the place. Like
I did for the actions, this will hopefully allow more easier api
additions and integrations.